refactor dashboard to subcomponents
